requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
bottleneck=1.3.7=py311hd7041d2_0
brotli=1.0.9=h2bbff1b_8
brotli-bin=1.0.9=h2bbff1b_8
brotli-python=1.0.9=py311hd77b12b_8
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.3.11=haa95532_0
certifi=2024.2.2=py311haa95532_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
contourpy=1.2.0=py311h59b6b97_0
cpuonly=2.0=0
cycler=0.11.0=pyhd3eb1b0_0
filelock=3.13.1=py311haa95532_0
fonttools=4.51.0=py311h2bbff1b_0
freetype=2.12.1=ha860e81_0
gmpy2=2.1.2=py311h7f96b67_0
icu=73.1=h6c2663c_0
idna=3.7=py311haa95532_0
intel-openmp=2021.4.0=haa95532_3556
jinja2=3.1.3=py311haa95532_0
jpeg=9e=h2bbff1b_1
kiwisolver=1.4.4=py311hd77b12b_0
krb5=1.20.1=h5b6d351_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_8
libbrotlidec=1.0.9=h2bbff1b_8
libbrotlienc=1.0.9=h2bbff1b_8
libclang=14.0.6=default_hb5a9fac_1
libclang13=14.0.6=default_h8e68704_1
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_1
libjpeg-turbo=2.0.0=h196d8e1_0
libpng=1.6.39=h8cc25b3_0
libpq=12.17=h906ac69_0
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libwebp-base=1.3.2=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_1
markupsafe=2.1.3=py311h2bbff1b_0
matplotlib=3.8.4=py311haa95532_0
matplotlib-base=3.8.4=py311hf62ec03_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py311h2bbff1b_0
mkl_fft=1.3.1=py311h743a336_0
mkl_random=1.2.2=py311heda8569_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpld3=0.5.9=py311haa95532_0
mpmath=1.3.0=py311haa95532_0
networkx=3.1=py311haa95532_0
numexpr=2.8.4=py311hffd1eac_0
numpy=1.24.3=py311ha6a8073_0
numpy-base=1.24.3=py311h877bb12_0
openjpeg=2.4.0=h4fc8c34_0
openssl=3.0.13=h2bbff1b_2
packaging=23.2=py311haa95532_0
pandas=2.2.1=py311hea22821_0
pillow=10.3.0=py311h2bbff1b_0
pip=24.0=py311haa95532_0
ply=3.11=py311haa95532_0
pyparsing=3.0.9=py311haa95532_0
pyqt=5.15.10=py311hd77b12b_0
pyqt5-sip=12.13.0=py311h2bbff1b_0
pysocks=1.7.1=py311haa95532_0
python=3.11.9=he1021f5_0
python-dateutil=2.9.0post0=py311haa95532_2
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=2.3.0=py3.11_cpu_0
pytorch-mutex=1.0=cpu
pytz=2024.1=py311haa95532_0
pyyaml=6.0.1=py311h2bbff1b_0
qt-main=5.15.2=h19c9488_10
requests=2.31.0=py311haa95532_1
setuptools=69.5.1=py311haa95532_0
sip=6.7.12=py311hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.45.3=h2bbff1b_0
sympy=1.12=py311haa95532_0
tk=8.6.14=h0416ee5_0
torchvision=0.18.0=py311_cpu
tornado=6.3.3=py311h2bbff1b_0
typing_extensions=4.11.0=py311haa95532_0
tzdata=2024a=h04d1e81_0
unicodedata2=15.1.0=py311h2bbff1b_0
urllib3=2.2.1=py311haa95532_0
vc=14.2=h2eaa2aa_1
vs2015_runtime=14.29.30133=h43f2093_3
wheel=0.43.0=py311haa95532_0
win_inet_pton=1.1.0=py311haa95532_0
xz=5.4.6=h8cc25b3_1
yaml=0.2.5=he774522_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.5=hd43e919_2
